+++ src/clock.h
@@ -1,8 +1,10 @@
 #ifndef CLOCK_H
 #define CLOCK_H
 
-void clock_set_date(int month, int day, int year);
-void clock_set_time(int h24, int min, int sec);
+#include <stdint.h>
+
+void clock_set_date(uint16_t month, uint16_t day, uint16_t year);
+void clock_set_time(uint16_t h24, uint16_t min, uint16_t sec);
 void clock_date_str(char *buffer);
 void clock_time_str(char *buffer);
 long clock_ticks();
